Split tests into unit and integration tests
To achieve this, we had to make ssl_checker into a proper module by giving it an `__init__.py` file. We had to do the same with the subdirectories `integration` and `unit`.

In order for the tests to work, `pytest` should be run as a python module from the root directory, i.e. `python -m pytest` - this runs it in the context of the root dir, and thus the tests can find the modules they need to import.